net.solarnetwork.common.internal.test/src/net/solarnetwork/codec/test/JodaDateTimeEpochSerializerTests.java
/* ==================================================================
 * JodaDateTimeEpochSerializerTests.java - 6/11/2019 9:19:54 am
 * 
 * Copyright 2019 SolarNetwork.net Dev Team
 * 
 * This program is free software; you can redistribute it and/or 
 * modify it under the terms of the GNU General Public License as 
 * published by the Free Software Foundation; either version 2 of 
 * the License, or (at your option) any later version.
 * 
 * This program is distributed in the hope that it will be useful, 
 * but WITHOUT ANY WARRANTY; without even the implied warranty of 
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
 * General Public License for more details.
 * 
 * You should have received a copy of the GNU General Public License 
 * along with this program; if not, write to the Free Software 
 * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 
 * 02111-1307 USA
 * ==================================================================
 */

package net.solarnetwork.codec.test;

import static org.hamcrest.Matchers.equalTo;
import static org.junit.Assert.assertThat;
import org.joda.time.DateTime;
import org.junit.Before;
import org.junit.Test;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.module.SimpleModule;
import net.solarnetwork.codec.JodaDateTimeEpochSerializer;

/**
 * Test cases for the {@link JodaDateTimeEpochSerializer} class.
 * 
 * @author matt
 * @version 1.0
 */
public class JodaDateTimeEpochSerializerTests {

	private ObjectMapper mapper;

	private ObjectMapper createObjectMapper() {
		ObjectMapper m = new ObjectMapper();
		SimpleModule mod = new SimpleModule("Test");
		mod.addSerializer(DateTime.class, new JodaDateTimeEpochSerializer());
		m.registerModule(mod);
		return m;
	}

	@Before
	public void setup() {
		mapper = createObjectMapper();
	}

	@Test
	public void serialize() throws Exception {
		DateTime now = new DateTime();
		String result = mapper.writeValueAsString(now);
		assertThat("Serialized date", result, equalTo(String.valueOf(now.getMillis())));
	}

}
